requirements.txt
appdirs==1.4.4
asgiref==3.4.1
astroid==2.6.5
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
black==21.7b0
cffi==1.15.0
channels==3.0.4
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pylint==2.9.6
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==21.0.0
pytz==2021.1
regex==2021.7.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.1
whitenoise==5.3.0
wrapt==1.12.1
zope.interface==5.4.0
